+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   gRPC.
 - The `yarpctest.FakeOutbound` can now send requests to a `transport.Router`.
   This allows end to end testing with a client and server in memory.
+- All outbounds now implement `Name` and a new `transport.Namer` interface.
+  This will allow outbound observability middleware to carry the transport name
+  in metrics properly.
 ### Changed
 - This change reduces the API surface of the peer list implementations to
   remove a previously public embedded type and replace it with implementations